minor fixes

+++ CapstonProjectBE/Controllers/PlatformController.cs
@@ -146,11 +146,11 @@ namespace CapstonProjectBE.Controllers
         public async Task<IActionResult> RemoveProjectPlatform(ProjectPlatformDTO projectPlatformDTO)
         {
             var user = await _authenService.GetUserByTokenAsync(HttpContext.User);
-            //var check = await _authenService.CheckIfUserHasPermissionsToUpdateOrDeleteByProjectId(projectPlatformDTO.ProjectId, user);
-            //if (check != null)
-            //{
-            //    return check;
-            //}
+            var check = await _authenService.CheckIfUserHasPermissionsToUpdateOrDeleteByProjectId(projectPlatformDTO.ProjectId, user);
+            if (check != null)
+            {
+                return check;
+            }
             var result = await _platformService.RemoveProjectPlatform(projectPlatformDTO);
             if (!result.Success)
             {
